Updated the image to properly incorporate msconvert
How it works: Wine is installed as root when building the image and only 
the owner of the wine directory can run wine. However, since running as 
a non-root user is the whole idea of singularity, I needed another 
approach. Setting chown did not work, since the user will never be owner 
of the directory and the user is not present in the singularity image 
(and duplicating the user into the image requires root).

Solution: Create a directory with links to all files from wine. This 
requires full access, but that means all the links will be owned by you, 
thus you can run msconvert