Added comments and fixed a problem which could have drastic consequences
The problem is detailed in the comments:

Quandenser needs the file_list with paths to the files.
Problem: We do not know the mzML file location during work and writing 
large files to publishDir is slow,
so there might be a possibility that some strange issues with incomplete 
mzML files being inputted to quandenser.
The solution is to have two publishdirs, one where the true files are 
being inserted for future use,
while we point in the file_list to the files in the work directory with 
symlinks.
The symlinks are fast to write + they point to complete files + we know 
the symlinks location --> fixed :)