See details, many changes
Fixed so ui can be run from inside the container. It is called from 
--app quandenser_ui

Removed wildcards, to be able to use singularity 3.0

Fixed some compatibility problems with singularity 3.0 and write 
permissions (ex read-only). The linking files are now on TMP with a 
randomly generated filename. This should work on uppmax as well (it 
didnt before)