Important change. Please read comment
config location changed from /var/tmp/quandenser_pipeline_$USER to
/home/$USER/.quandenser_pipeline

Why?

In clusters, there are usually several login nodes. Nextflow is run from
one of the login nodes and will use that node's environment. When
logging in to say rackham, there are 3 possible login nodes that you
might be logged in at. Each node has it's own /var/tmp folder, meaning
if you logged in at rackham1 and then to rackham2, the settings and
everything like that would be different. However, the home directory
is shared between nodes which makes any files in home persist.

Another problem with login nodes are that if you login to rackham1 and
run a script there, it will not be accessible to rackham2. This prevents
quandenser-pipeline from interacting with that process. The solution is:

1: Login to the same node that nextflow was run at

2: When the user tries to kill a process on a different node, alert the
user of the problem