Quandenser-pipeline v0.04
Changelog:

Added MSconvert to the pipeline

Added option to parallelize msconvert and quandenser, with limits on how
many processes you can do simultaneously

Added a modified quandenser, which can handle parallelization

Improved the channels to prevent name collisions

Config location changed from /var/tmp/quandenser_pipeline_$USER to
/home/$USER/.quandenser_pipeline The reason why is that in clusters,
there are usually several login nodes. Nextflow is started from
one of the login nodes and will use that node's environment. When
logging in to say rackham, there are 3 possible login nodes that you
might be logged in at. Each node has it's own /var/tmp folder, meaning
if you logged in at rackham1 and then to rackham2, the settings and
everything like that would be different. However, the home directory
is shared between nodes which makes any files in home persist.
Another problem with login nodes are that if you login to rackham1 and
run a script there, it will not be accessible to rackham2. This prevents
quandenser-pipeline from interacting with that process.

Added tooltips to some options

Fixed problem with links with msconvert

Added option to resume from a previous directory (aka all previous files
are loaded in when running the next run). This will be useful if the
workflow crashes or you run out of calculation time on the cluster, so
you don't lose all your data.

Added so you can get an email when the pipeline has finished. Only works
with gmail accounts (this includes scilifelab emails). If you have 2
factor authorization, go to:
https://security.google.com/settings/security/apppasswords
to generate a token, to use instead of a password.
Note: Since all files you create are only readable by you (chmod 700),
your password is secure in the nf.config file

Added "labels" to the output folder. The purpose is that if you want to
run many analyzes in the same folder, you add a label (ex "ecoli_gut"),
quandenser-pipeline will create a new directory for you (ecoli_gut_1)
that will prevent outputs from overwriting one another. All runs will
use the same work directory, without collisions between them.

Added right-click on label field on tab1 to automatically assign an
unique label for each file that is unlabeled

Improved robustness with the pipe script, which prevents crashes on
clusters. Note that some issues are still prevalent, please go to Readme
in the github page for known issues.

Fixed some typos that crashed the nextflow pipeline in some cases

Fixed a lot of weird bugs in the UI. All parameters should now write
properly to their assigned location

Improved defaults for slurm cluster computation