Quandenser-pipeline v0.05
CHANGELOG

Changes in building the image:

The quandenser version is prepackaged in the "dependencies" directory,
which uses the modified quandenser on this branch:
https://github.com/statisticalbiotechnology/quandenser/tree/quandenser-pipeline

The triqler version is now using another branch and not the pip version.
The modified version includes outputting raw intensities, and will be
used to add additional post processing steps
https://github.com/statisticalbiotechnology/triqler/tree/quandenser-pipeline

Changes in the image:

Major:

Increased parallelization functionality by saving and loading
parameters. For a data set of about 80 files, this gives 7.5 hours
faster calculation time (!) There are still many more areas where you
can speed up the processing, but for now it is enough. This MIGHT yield
strange bugs when running, which I am not working to check.

Fixed bug where path to converted files were wrong,
which made RAW files not able to process due to the wrong paths.
(Sorry about that)

Fixed bug where maracluster freezes if you run large cohorts,
which is now included in the new version of quandenser

Minor:

Auto-labeling is now using numbers instead of characters
(related to issue)
https://github.com/statisticalbiotechnology/quandenser-pipeline/issues/4

Improved check for crashes during start of run. It now waits 2 seconds
to check if nextflow managed to run or not.

Increased limit for spinboxes, so you are now not limited to 99

Changed "" to '' in nf.config so you can use " in additional arguments

Added basic copy and paste on the file table
(This is not saved to the clipboard at the moment, since Qclipboard
doesnt seem to work. The copy and paste are used internally). This will
be improved in the future

Enter when end of file choosing table creates new row

Fixed minor bugs with empty rows (they should not be happening anymore)

Added right-click on the "file" column to remove empty rows

The new quandenser file packaged comes from a different branch of
quandenser, to minimize the work to repack the file each time

Added option to output raw intensities from triqler (WIP)

Added defaults slightly to more accurately be compatible with HPC
running time