Quandenser-pipeline v0.06-beta
### Changes in shell script

* Added so the shell script checks for newer versions of the pipeline. 
This will overwrite the older image and download the newest version of 
the shell script. This should be compatible with all releases in the 
future

* Improved help description

* Added license of Proteowizard

* Added short commands to every option

### Changes in building the image:

* Added "zip" and "tkinter for python3.6" in the singularity image. Zip 
will allow for triqler sending files via email and tkinter will prevent 
python leakages to the host.

### Changes in the image:

**Major changes:**

* You can now send the triqler files with email when the pipeline is 
done! Just set "send files" in the advanced tab to "true".

* Changed so msconvert can output any format, not just mzML

* Changed so python uses only the image modules and will no longer 
"leak" 
into the host computer, i.e use the hosts python modules will not be 
loaded if there both the host and the image has the same module 
installed.

* Added option to ignore tree parallelization, so you only parallelize 
only the first part. Disabling the tree parallelization should prevent 
bugs related to issue #15 . Quandenser has also been slightly modified 
to make this work


**Minor changes:**

* Improved defaults when running on cluster. Also condensed the in the 
workflow, putting the cpu spinboxes besides the time boxes

* Changed back ' to " in nf.config, since mixing the quotes is not 
compatible with some commands. The user can only use ' in the additional 
commands

* Added some additional commands in the config parser and fixed some 
minor bugs.

* Removed a crash warning if the directory is not created within 2 
seconds. Sometimes, it is just slow. If that is the case, it will still 
say quandenser started

* Added experimental raw intensities and distributions output. These are 
a work in progress, and should not really be used, but they are 
available.

* Fixed limits for spinboxes related to issue #11 

* Improved how tooltips are loaded. This will not affect the users, but 
it will allow me to much easier document and add tooltips to the GUI :)

* Improved About tab in the GUI. A neat feature about the About tab is 
that it is a mini browser, so you can now click the links inside and 
read about the software used! (Requires open-gl enabled)

* The workflow has been improved in the edit workflow tab. It is now 
much clearer how the workflow works.  (Requires open-gl enabled)