IMPORTANT UPDATE ON .SH FILE:
Due to a [new update on singularity 
hub]((https://singularityhub.github.io/singularityhub-docs/2019/security-release/), 
the API get requests are limited in number, meaning the it will not be 
possible to determine version number by using the Singularity hub API, 
thus preventing updates. Therefore, the new .sh file will use github's 
api to fetch the version number. This means that both the tag name of 
each release in github and the version number in the image needs to be 
updated on each release. Before, only the version number in the image 
was needed to update.