Quandenser-pipeline v0.08-beta
### Changes in the shell script:

**Major changes:**

* Due to a [new update on singularity 
hub](https://singularityhub.github.io/singularityhub-docs/2019/security-release/), 
 the API get requests are limited, meaning the it will not be possible 
to determine version number by using the Singularity hub API, thus 
preventing updates from the shellscript. Therefore, the new .sh file 
will use github's api to fetch the version number instead. This means 
that both the tag name of  each release in github and the version number 
in the image needs to be  updated on each release while before, only the 
version number in the image was needed to update. **THIS MEANS THAT 
USERS BEFORE 0.08 NEEDS TO UPDATE TO THE NEWEST VERSION OF THE SHELL 
SCRIPT, SINCE THE OLD VERSION WILL NOT AUTOMATICALLY FIND UPDATES**. 
Please download the newest version of the shell script to fix this 
issue.

### Changes in building the image:

**Minor changes:**

* [New version of nextflow requires the user to run it in the 
background](https://github.com/nextflow-io/nextflow/issues/1221):
This breaks several things which will require some rework. With the new 
nextflow version, *run_quandenser.sh* will suspend all processes if the 
tag "-bf" is not added. Adding the "background" tag will break the 
ability to kill processes in the GUI, since the .sh script will no 
longer have nextflow as its child-process. Some functions used in the 
pipeline will be depreciated in future nextflow versions, which might 
cause further problems. Thus, the update for nextflow 19.01 --> 19.07 
causes a lot of issues, which might not be worth fixing at the moment. 
Nextflow version 19.01 will be the default, static version for now.

* Added some permission for the new quandenser version, so users can run 
it without any problems.

### Changes in the image:

**Major changes:**

* A new version of quandenser has been installed 
([v0.02](https://github.com/statisticalbiotechnology/quandenser/releases/tag/rel-0-02)) 
which includes updated versions of submodules and huge improvements to 
speed of parallelization and the stability of the processing.

* A new triqler version has been added 
([v0.03](https://github.com/statisticalbiotechnology/triqler/releases/tag/rel-0-03-0), 
which has several improvements, such as optimizations for very large 
datasets.

**Minor changes:**

* Removed the ability to choose max missing value directly from the GUI. 
The reason is that few users know what the the option does and that it 
should be changed depending on the amount of files available. The 
default value in quandenser takes this into account, but is overwritten 
by the GUI value. The value can still be changed by adding 
"--max-missing" <value> into quandenser's optional parameters. This 
should prevent confusion of new users, but still allow users to change 
the values to their liking.

* Removed parallel step 5, by merging 4 and 5. The last step is not 
necessary and only adds to the total calculation time.

* Fixed some bugs in the nextflow pipeline when running quandenser v0.02

* Fixed some bugs where elements of the GUI are hidden if the window 
size is too small.