Quandenser-pipeline v0.082
### Changes in building the image:

**Minor changes:**

* 
[command_wrapper.py](https://github.com/statisticalbiotechnology/quandenser-pipeline/blob/master/dependencies/command_wrapper.py) 
added to the image. The purpose is to fix some common errors found in 
the pipeline.

### Changes in the shell script:

**Minor changes:**

* Added messages when the user enables some options

* The installed singularity version by the shell script is now fixed on 
singularity version v3.2.1, since the latest version v3.4.0 has some 
restrictions on what you can interact with on the host computer. This 
includes the "running jobs" tab and the button which can kill the 
processes. This means you would have to kill all processes manually, if 
you have v3.4.0 installed

### Changes in the image:

**Major changes:**

* Major improvements with parallel processing. See enhancement 
https://github.com/statisticalbiotechnology/quandenser-pipeline/issues/19. 
A positive side effect of this change is that the parallel processing is 
way more efficient with memory, so that means you can run larger data 
sets with the parallel option!

**Minor changes:**

* This is a "minor" change, but this change fixes 2 very important and 
"pipeline breaking" bugs, namely issue #3 and issue #23. MSconvert, 
quandenser_parallel_1 and quandenser_parallel_3 will now be executed 
through a "wrapper" script, which checks for known errors which causes 
the issue. The script is found 
[here](https://github.com/statisticalbiotechnology/quandenser-pipeline/blob/master/dependencies/command_wrapper.py). 
This should also prevent zombie processes from happening.

* You can now choose in tab 3 if you want to publish the files (aka copy 
them to the output directory from work directory). This is an useful 
option to have if you have limited storage space on the computer you are 
working on and don't want some of the "intermediate" files. Note that 
you still have to remove the work directory manually when you are done.

* The pipeline now allows one crash on msconvert, quandenser_parallel_1 
and quandenser_parallel_3, due to the known errors and command_wrapper. 
This also solves in most cases issue #22, where singularity fails to 
mount the image when running many parallel processes. It is sometimes 
rare, but if the pipeline fails to mount the image twice in a row on the 
same process, the pipeline will crash. This fix helps at least computers 
which has that issue from crashing 100% of the time, to a lot less than 
that.